add readme and cleanup

+++ pages/api/posts/[id].js
@@ -1,22 +1,33 @@
 import prisma from "../../../lib/prisma"
+import { getSession } from "next-auth/client"
 
-export default async function handle(req, res) {
+export default async (req, res) => {
   const id = req.query.id
 
   let post
 
   if (req.method === "DELETE") {
     // DESTROY
-    post = await prisma.post.delete({
-      where: { id: Number(id) },
-    })
+    const session = await getSession({ req })
+    if (session) {
+      post = await prisma.post.delete({
+        where: { id: Number(id) },
+      })
+    } else {
+      res.status(401)
+    }
   } else if (req.method === "PUT") {
     // UPDATE
-    const { title, content } = req.body
-    post = await prisma.post.update({
-      where: { id: Number(id) },
-      data: { title: title, content: content },
-    })
+    const session = await getSession({ req })
+    if (session) {
+      const { title, content } = req.body
+      post = await prisma.post.update({
+        where: { id: Number(id) },
+        data: { title: title, content: content },
+      })
+    } else {
+      res.status(401)
+    }
   } else {
     // SHOW
     post = await prisma.post.findFirst({
